user-service/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0